@@ -40,18 +40,14 @@ class VirtualEnvHandler(object):
     def activate_virtualenv(self):
         """Activates the virtualenv at `VIRTUALENV_PATH`."""
         logging.info("Activating virtualenv at: `{}`".format(self._virtualenv_path))
-        subprocess.check_output(
-            "source {}/bin/activate".format(self._virtualenv_path), shell=True
-        )
+        subprocess.check_output("source {}/bin/activate".format(self._virtualenv_path), shell=True)
         self._is_activated = True
 
     def create_virtualenv(self):
         """Clears the folder and creates a new virtualenv at
         `self._virtualenv_path`."""
         if self.virtualenv_exists():
-            logging.ERROR(
-                "Virtualenv already exists at: `{0}`".format(self._virtualenv_path)
-            )
+            logging.ERROR("Virtualenv already exists at: `{0}`".format(self._virtualenv_path))
             return
         logging.info("Creating virtualenv at: `{}`".format(self._virtualenv_path))
         venv.create(self._virtualenv_path, clear=True, with_pip=True)
@@ -126,9 +122,7 @@ class Suite(object):
 
         This function must be implemented in all child classes of `Suite.`
         """
-        raise NotImplementedError(
-            "Must implement `register_benchmarks` to run benchmark suite."
-        )
+        raise NotImplementedError("Must implement `register_benchmarks` to run benchmark suite.")
 
     def before_all(self):
         pass
@@ -225,11 +219,7 @@ class Runner(object):
         """Wrap the benchmark `func` with timing code and run for n
         `ITERATIONS`, returning a result row that can be fed into Perspective.
         """
-        overall_result = {
-            k.replace("__BENCH__", ""): v
-            for (k, v) in func.__dict__.items()
-            if "__BENCH__" in k
-        }
+        overall_result = {k.replace("__BENCH__", ""): v for (k, v) in func.__dict__.items() if "__BENCH__" in k}
 
         result = timeit(func, number=Runner.ITERATIONS) / Runner.ITERATIONS
         overall_result["__TIME__"] = result
