@@ -1,6 +1,6 @@
 ################################################################################
 #
-# Copyright (c) 2022, the Perspective Authors.
+# Copyright (c) 2019, the Perspective Authors.
 #
 # This file is part of the Perspective library, distributed under the terms of
 # the Apache License 2.0.  The full license can be found in the LICENSE file.
